+++ include/GameAnalytics/GameAnalytics.h
@@ -103,6 +103,8 @@ namespace gameanalytics
          static void endSession();
 
          static std::string getRemoteConfigsValueAsString(std::string const& key, std::string const& defaultValue = "");
+         static std::string getRemoteConfigsValueAsJson(std::string const& key);
+
 
          static bool        isRemoteConfigsReady();
          static void        addRemoteConfigsListener(const std::shared_ptr<IRemoteConfigsListener> &listener);
@@ -116,6 +118,9 @@ namespace gameanalytics
          static std::string getABTestingId();
          static std::string getABTestingVariantId();
 
+         static int64_t getElapsedSessionTime();
+         static int64_t getElapsedTimeFromAllSessions();
+
          // game state changes
          // will affect how session is started / ended
          static void onResume();
@@ -125,10 +130,9 @@ namespace gameanalytics
          static bool isThreadEnding();
 
      private:
-         static bool _endThread;
 
-        static bool isSdkReady(bool needsInitialized);
-        static bool isSdkReady(bool needsInitialized, bool warn);
-        static bool isSdkReady(bool needsInitialized, bool warn, std::string const& message);
+        static bool _endThread;
+        static bool isSdkReady(bool needsInitialized, bool warn = true, std::string const& message = "");
     };
+
 } // namespace gameanalytics